dom/apps/tests/test_bug_1168300.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1168300
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1168300</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript;version=1.7">

  /** Test for Bug 1168300 **/

  SimpleTest.waitForExplicitFinish();

  var url = "http://test1.example.com/tests/dom/apps/tests/file_app.sjs?apptype=hosted&getmanifest=true";

  var gGenerator = runTest();

  function go() {
    SpecialPowers.pushPermissions(
      [{ "type": "webapps-manage", "allow": 1, "context": document }],
      function() { gGenerator.next() });
  }

  function continueTest() {
    try {
      gGenerator.next();
    } catch (e if e instanceof StopIteration) {
      SimpleTest.finish();
    }
  }

  function mozAppsError() {
    ok(false, "mozApps error: " + this.error.name);
    SimpleTest.finish();
  }

  function runTest() {
    // Set up.
    SpecialPowers.setAllAppsLaunchable(true);
    SpecialPowers.autoConfirmAppInstall(continueTest);
    yield undefined;

    SpecialPowers.autoConfirmAppUninstall(continueTest);
    yield undefined;

    let request = navigator.mozApps.install(url);
    request.onerror = mozAppsError;
    request.onsuccess = continueTest;
    yield undefined;
    let app = request.result;

    let _topic = "clear-cookiejar-data";
    let observer = new Observer(_topic);
    observer.onobserve = function(subject, topic, data, count) {
      ok(topic == _topic, "unknown topic " + topic);

      let cookieJar = data;
      ok(cookieJar, "params.cookieJar should have value for an app.");

      // should receive 2 notifications when an app is uninstalled.
      if (count == 2) {
        ok(true);
        observer.shutdown();
        continueTest();
      }
    };

    request = navigator.mozApps.mgmt.uninstall(app);
    request.onsuccess = function() {
      // wait for clear-cookiejar-data event.
    };
    request.onerror = mozAppsError;
    yield undefined;
  }

  function Observer(topic) {
    this.topic = topic;
    SpecialPowers.addObserver(this, topic, false);
  }
  Observer.prototype = {
    topic: null,

    count: 1,

    callback: null,

    observe: function(subject, topic, data) {
      if (this.callback) this.callback(subject, topic, data, this.count++);
    },

    shutdown: function() {
      SpecialPowers.removeObserver(this, this.topic);
    },

    set onobserve (callback) {
      this.callback = callback;
    }
  };

  SimpleTest.waitForExplicitFinish();
  </script>
</head>
<body onload="go()">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1168300">Mozilla Bug 1168300</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
